# Conflicts:
#	DDI/Services/DDI.Services.csproj

+++ DDI/DDI.Data/SQLUtilities.cs
@@ -1,6 +1,7 @@
 ﻿using DDI.Shared;
 using System.Data.Entity;
 using System.Linq;
+using System;
 
 namespace DDI.Data
 {
@@ -19,7 +20,7 @@ namespace DDI.Data
         /// <summary>
         /// Set the next value for a database sequence.
         /// </summary>
-        public void SetNextSequenceValue(string sequenceName, int newValue)
+        public void SetNextSequenceValue(string sequenceName, Int64 newValue)
         {
             _context.Database.ExecuteSqlCommand($"ALTER SEQUENCE {sequenceName} RESTART WITH {newValue};");
         }
@@ -27,9 +28,9 @@ namespace DDI.Data
         /// <summary>
         /// Get the next value for a database sequence.
         /// </summary>
-        public int GetNextSequenceValue(string sequenceName)
+        public Int64 GetNextSequenceValue(string sequenceName)
         {
-            return _context.Database.SqlQuery<int>($"SELECT NEXT VALUE FOR {sequenceName};").FirstOrDefault();
+            return _context.Database.SqlQuery<Int64>($"SELECT NEXT VALUE FOR {sequenceName};").FirstOrDefault();
         }
         
     }
